requirements/requirements.txt
numpy>=1.26,<3
pandas>=2.2,<3
scikit-learn>=1.4,<2
feature-engine>=1.8,<2
pydantic>=2.6,<3
PyYAML>=6,<7
joblib>=1.3,<2
